requirements.txt
ml-dtypes>=0.5.3
pillow>=11.3.0
appdirs
astropy>=6.0.1
astropy-iers-data>=0.2025.9.1.0.42.11
customtkinter>=5.2.2
imageio>=2.37.0
minio>=7.2.16
numpy>=2.0.2,<=2.2.6
opencv-python-headless==4.11.0.86
PyKrige>=1.7.2
requests>=2.32.5
scikit-image>=0.24.0
scipy>=1.13.1
xisf>=0.9.5